ref(project-creation): Extract shared SCM platform resolution hook
ScmPlatformFeaturesCore and ScmFeatureSelectionPanel each re-derived the
resolved platform from the detection query with identical reduce logic, kept in
sync only by a comment. Move that derivation into useScmResolvedPlatform so the
invariant lives in one place. Both call sites still call it independently and
rely on the deduped detection query rather than lifted shared state, preserving
the re-derivation design.

Also replace the repeated [ProductSolution.ERROR_MONITORING] default with a
shared DEFAULT_SCM_FEATURES constant.
